feat: rewrite codegen, remove old code and prepare libpg_query parsing
