+++ crates/codegen/src/syntax_kind.rs
@@ -1,13 +1,10 @@
 use std::collections::HashSet;
 
-use pg_query_proto_parser::{Node, ProtoParser, Token};
+use pg_query_proto_parser::{Node, ProtoFile, Token};
 use proc_macro2::{Ident, Literal};
 use quote::{format_ident, quote};
 
-pub fn syntax_kind_mod(_item: proc_macro2::TokenStream) -> proc_macro2::TokenStream {
-    let parser = ProtoParser::new("libpg_query/protobuf/pg_query.proto");
-    let proto_file = parser.parse();
-
+pub fn syntax_kind_mod(proto_file: &ProtoFile) -> proc_macro2::TokenStream {
     let custom_node_names = custom_node_names();
     let custom_node_identifiers = custom_node_identifiers(&custom_node_names);
 
@@ -15,9 +12,6 @@ pub fn syntax_kind_mod(_item: proc_macro2::TokenStream) -> proc_macro2::TokenStr
 
     let token_identifiers = token_identifiers(&proto_file.tokens);
     let token_value_literals = token_value_literals(&proto_file.tokens);
-
-    let syntax_kind_impl =
-        syntax_kind_impl(&node_identifiers, &token_identifiers, &token_value_literals);
 
     let syntax_kind_from_impl =
         syntax_kind_from_impl(&node_identifiers, &token_identifiers, &token_value_literals);
@@ -29,9 +23,6 @@ pub fn syntax_kind_mod(_item: proc_macro2::TokenStream) -> proc_macro2::TokenStr
     let unique_enum_variants = enum_variants.into_iter().collect::<Vec<_>>();
 
     quote! {
-        use cstree::Syntax;
-        use pg_query::{protobuf::ScanToken, NodeEnum, NodeRef};
-
         /// An u32 enum of all valid syntax elements (nodes and tokens) of the postgres
         /// sql dialect, and a few custom ones that are not parsed by pg_query.rs, such
         /// as `Whitespace`.
@@ -42,7 +33,6 @@ pub fn syntax_kind_mod(_item: proc_macro2::TokenStream) -> proc_macro2::TokenStr
         }
 
         #syntax_kind_from_impl
-        #syntax_kind_impl
     }
 }
 
@@ -86,23 +76,6 @@ fn token_value_literals(tokens: &[Token]) -> Vec<Literal> {
         .collect()
 }
 
-fn syntax_kind_impl(
-    node_identifiers: &[Ident],
-    token_identifiers: &[Ident],
-    token_value_literals: &[Literal],
-) -> proc_macro2::TokenStream {
-    let new_from_pg_query_node_fn = new_from_pg_query_node_fn(node_identifiers);
-    let new_from_pg_query_token_fn =
-        new_from_pg_query_token_fn(token_identifiers, token_value_literals);
-    quote! {
-        impl SyntaxKind {
-            #new_from_pg_query_node_fn
-
-            #new_from_pg_query_token_fn
-        }
-    }
-}
-
 fn syntax_kind_from_impl(
     node_identifiers: &[Ident],
     token_identifiers: &[Ident],
@@ -116,6 +89,16 @@ fn syntax_kind_from_impl(
                     #(NodeEnum::#node_identifiers(_) => SyntaxKind::#node_identifiers),*
                 }
             }
+
+        }
+
+        impl From<Token> for SyntaxKind {
+            fn from(token: Token) -> SyntaxKind {
+                match i32::from(token) {
+                    #(#token_value_literals => SyntaxKind::#token_identifiers),*,
+                    _ => panic!("Unknown token: {:?}", token),
+                }
+            }
         }
 
         impl From<&ScanToken> for SyntaxKind {
@@ -124,34 +107,6 @@ fn syntax_kind_from_impl(
                     #(#token_value_literals => SyntaxKind::#token_identifiers),*,
                     _ => panic!("Unknown token: {:?}", token.token),
                 }
-            }
-        }
-    }
-}
-
-fn new_from_pg_query_node_fn(node_identifiers: &[Ident]) -> proc_macro2::TokenStream {
-    quote! {
-
-        /// TODO: drop this
-        pub fn new_from_pg_query_node(node: &NodeEnum) -> Self {
-            match node {
-                #(NodeEnum::#node_identifiers(_) => SyntaxKind::#node_identifiers),*
-            }
-        }
-    }
-}
-
-fn new_from_pg_query_token_fn(
-    token_identifiers: &[Ident],
-    token_value_literals: &[Literal],
-) -> proc_macro2::TokenStream {
-    quote! {
-
-        /// Converts a `pg_query` token to a `SyntaxKind`
-        pub fn new_from_pg_query_token(token: &ScanToken) -> Self {
-            match token.token {
-                #(#token_value_literals => SyntaxKind::#token_identifiers),*,
-                _ => panic!("Unknown token: {:?}", token.token),
             }
         }
     }
