--- crates/parser/src/get_child_token_range_codegen.rs
@@ -1,3 +1,0 @@
-use codegen::get_child_token_range;
-
-get_child_token_range!();
