--- crates/parser/src/get_location_codegen.rs
@@ -1,3 +1,0 @@
-use codegen::get_location;
-
-get_location!();
