--- crates/parser/src/syntax_kind_codegen.rs
@@ -1,3 +1,0 @@
-use codegen::syntax_kind;
-
-syntax_kind!();
